+++ extension/src/ALObject/ALPageField.ts
@@ -1,6 +1,7 @@
+import { ALDataType } from "./ALDataType";
 import { ALPageControl } from "./ALPageControl";
 import { ALTableField } from "./ALTableField";
-import { ALControlType, ALPropertyType } from "./Enums";
+import { ALControlType, ALPropertyType, DataType } from "./Enums";
 
 export class ALPageField extends ALPageControl {
   public get caption(): string {
@@ -11,7 +12,14 @@ export class ALPageField extends ALPageControl {
 
     // Check table for caption
     const field = this.getSourceTableField();
-    return field ? field.caption : "";
+    if (!field) {
+      return "";
+    }
+    if (field.caption !== "") {
+      return field.caption;
+    }
+    const sourceObject = this.getObject().getSourceObject();
+    return sourceObject && sourceObject.isSystemObject ? field.name : "";
   }
 
   public get readOnly(): boolean {
@@ -24,7 +32,33 @@ export class ALPageField extends ALPageControl {
 
     // Check table field
     const field = this.getSourceTableField();
-    return field ? !field.getProperty(ALPropertyType.editable, true) : false;
+    return field
+      ? !field.getProperty(ALPropertyType.editable, true) || field.isSystemField
+      : false;
+  }
+
+  public get dataType(): ALDataType | undefined {
+    // Check table field
+    const field = this.getSourceTableField();
+    if (field) {
+      return field.dataType;
+    }
+    // Check global variables
+    let variableName = this.value;
+    if (this.value.includes("[")) {
+      variableName = this.value.slice(0, this.value.indexOf("["));
+      const variable = this.getObject().variables.find(
+        (f) => f.name === variableName
+      );
+      return variable
+        ? new ALDataType(variable.type.subtype as DataType)
+        : undefined;
+    } else {
+      const variable = this.getObject().variables.find(
+        (f) => f.name === variableName
+      );
+      return variable ? variable.type : undefined;
+    }
   }
 
   private getSourceTableField(): ALTableField | undefined {
@@ -37,7 +71,16 @@ export class ALPageField extends ALPageControl {
     const fields = allControls.filter(
       (x) => x.type === ALControlType.tableField
     );
-    const field = fields.find((x) => x.name === this.value);
+    sourceObject.extensionObjects.forEach((obj) =>
+      fields.push(
+        ...obj
+          .getAllControls()
+          .filter((x) => x.type === ALControlType.tableField)
+      )
+    );
+    const field = fields.find(
+      (x) => x.name.toLowerCase() === this.value.toLowerCase()
+    );
     return field as ALTableField;
   }
 }