+++ extension/src/SymbolReference/SymbolReferenceReader.ts
@@ -18,6 +18,7 @@ import { AppPackage } from "./types/AppPackage";
 import { symbolReferenceCache } from "./SymbolReferenceCache";
 import { ALPageField } from "../ALObject/ALPageField";
 import { ALPagePart } from "../ALObject/ALPagePart";
+import { addSystemFields } from "../ALObject/ALParser";
 
 export function getObjectsFromAppFile(appFilePath: string): AppPackage {
   const appIdentifier = AppPackage.appIdentifierFromFilename(appFilePath);
@@ -90,17 +91,14 @@ function tableToObject(table: TableDefinition): ALObject {
     addProperty(prop, obj);
   });
   table.Fields?.forEach((field) => {
-    const alField = new ALTableField(
-      ALControlType.tableField,
-      field.Id as number,
-      field.Name,
-      field.TypeDefinition.Name
-    );
+    const alField = ALTableField.fromFieldDefinition(field);
+    alField.parent = obj;
     field.Properties?.forEach((prop) => {
       addProperty(prop, alField);
     });
     obj.controls.push(alField);
   });
+  addSystemFields(obj);
   return obj;
 }